Make `oep2 report` run against the local repo by default
