+++ dev-requirements.in
@@ -1,8 +1,8 @@
 pytest
 pytest-mock
-pytest-xdist
 responses
 pip-tools
 pep8
 pylint
 edx-lint
+
