+++ edx_repo_tools/oep2/oep2-report.ini
@@ -1,0 +1,4 @@
+[pytest]
+python_files=check_*.py
+python_classes=Check
+python_functions=check_*
