+++ requirements.txt
@@ -5,6 +5,7 @@
 #    pip-compile --output-file requirements.txt requirements.in
 #
 ansicolors==1.0.2
+apipkg==1.4               # via execnet
 appdirs==1.4.0
 attrs==16.0.0             # via service-identity
 CacheControl==0.10.7
@@ -14,6 +15,7 @@ cryptography==1.4         # via pyopenssl
 cssselect==0.9.2          # via parsel, scrapy
 docutils==0.12            # via statistics
 enum34==1.1.6             # via cryptography
+execnet==1.4.1            # via pytest-xdist
 Flask==0.10.1
 gitdb==0.6.4              # via gitpython
 github3.py==0.9.5
@@ -34,12 +36,13 @@ ordereddict==1.1          # via python-dotenv
 parsel==1.0.3             # via scrapy
 path.py==8.2.1
 psycopg2==2.5.4
-py==1.4.31                # via pytest
+py==1.4.31                # via pytest, pytest-xdist
 pyasn1-modules==0.0.8     # via service-identity
 pyasn1==0.1.9             # via cryptography, pyasn1-modules, service-identity
 pycparser==2.14           # via cffi
 PyDispatcher==2.0.5       # via scrapy
 pyOpenSSL==16.0.0         # via scrapy, service-identity
+pytest-xdist==1.15.0
 pytest==2.9.2
 python-dateutil==2.5.3
 python-dotenv==0.5.1
