--- setup.cfg
@@ -1,2 +1,0 @@
-[pytest]
-addopts=--ignore edx_repo_tools/oep2/tests
